+++ Examples/Executor/Executor.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
 using QuickFix;
 using QuickFix.Fields;
@@ -48,11 +45,11 @@ namespace Executor
             Price price = new Price(DEFAULT_MARKET_PRICE);
             ClOrdID clOrdID = n.ClOrdID;
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -67,10 +64,10 @@ namespace Executor
                 symbol,
                 side,
                 orderQty,
-                new LastShares(orderQty.getValue()),
-                new LastPx(price.getValue()),
-                new CumQty(orderQty.getValue()),
-                new AvgPx(price.getValue()));
+                new LastShares(orderQty.Value),
+                new LastPx(price.Value),
+                new CumQty(orderQty.Value),
+                new AvgPx(price.Value));
 
             exReport.Set(clOrdID);
 
@@ -101,11 +98,11 @@ namespace Executor
             Price price = new Price(DEFAULT_MARKET_PRICE);
             ClOrdID clOrdID = n.ClOrdID;
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -121,11 +118,11 @@ namespace Executor
                 symbol,
                 side,
                 orderQty,
-                new LastShares(orderQty.getValue()),
-                new LastPx(price.getValue()),
+                new LastShares(orderQty.Value),
+                new LastPx(price.Value),
                 new LeavesQty(0),
-                new CumQty(orderQty.getValue()),
-                new AvgPx(price.getValue()));
+                new CumQty(orderQty.Value),
+                new AvgPx(price.Value));
 
             exReport.Set(clOrdID);
 
@@ -156,11 +153,11 @@ namespace Executor
             ClOrdID clOrdID = n.ClOrdID;
             Price price = new Price(DEFAULT_MARKET_PRICE);
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -176,13 +173,13 @@ namespace Executor
                 symbol,
                 side,
                 new LeavesQty(0),
-                new CumQty(orderQty.getValue()),
-                new AvgPx(price.getValue()));
+                new CumQty(orderQty.Value),
+                new AvgPx(price.Value));
 
             exReport.Set(clOrdID);
             exReport.Set(orderQty);
-            exReport.Set(new LastShares(orderQty.getValue()));
-            exReport.Set(new LastPx(price.getValue()));
+            exReport.Set(new LastShares(orderQty.Value));
+            exReport.Set(new LastPx(price.Value));
 
             if (n.IsSetAccount())
                 exReport.SetField(n.Account);
@@ -211,11 +208,11 @@ namespace Executor
             Price price = new Price(DEFAULT_MARKET_PRICE);
             ClOrdID clOrdID = n.ClOrdID;
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -230,14 +227,14 @@ namespace Executor
                 symbol, // Shouldn't be here?
                 side,
                 new LeavesQty(0),
-                new CumQty(orderQty.getValue()),
-                new AvgPx(price.getValue()));
+                new CumQty(orderQty.Value),
+                new AvgPx(price.Value));
 
             exReport.Set(clOrdID);
             exReport.Set(symbol);
             exReport.Set(orderQty);
-            exReport.Set(new LastQty(orderQty.getValue()));
-            exReport.Set(new LastPx(price.getValue()));
+            exReport.Set(new LastQty(orderQty.Value));
+            exReport.Set(new LastPx(price.Value));
 
             if (n.IsSetAccount())
                 exReport.SetField(n.Account);
@@ -266,11 +263,11 @@ namespace Executor
             Price price = new Price(DEFAULT_MARKET_PRICE);
             ClOrdID clOrdID = n.ClOrdID;
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -285,14 +282,14 @@ namespace Executor
                 symbol, //shouldn't be here?
                 side,
                 new LeavesQty(0),
-                new CumQty(orderQty.getValue()),
-                new AvgPx(price.getValue()));
+                new CumQty(orderQty.Value),
+                new AvgPx(price.Value));
 
             exReport.Set(clOrdID);
             exReport.Set(symbol);
             exReport.Set(orderQty);
-            exReport.Set(new LastQty(orderQty.getValue()));
-            exReport.Set(new LastPx(price.getValue()));
+            exReport.Set(new LastQty(orderQty.Value));
+            exReport.Set(new LastPx(price.Value));
 
             if (n.IsSetAccount())
                 exReport.SetField(n.Account);
@@ -321,11 +318,11 @@ namespace Executor
             Price price = new Price(DEFAULT_MARKET_PRICE);
             ClOrdID clOrdID = n.ClOrdID;
 
-            switch (ordType.getValue())
+            switch (ordType.Value)
             {
                 case OrdType.LIMIT:
                     price = n.Price;
-                    if (price.Obj == 0)
+                    if (price.Value == 0)
                         throw new IncorrectTagValue(price.Tag);
                     break;
                 case OrdType.MARKET: break;
@@ -339,14 +336,14 @@ namespace Executor
                 new OrdStatus(OrdStatus.FILLED),
                 side,
                 new LeavesQty(0),
-                new CumQty(orderQty.getValue()));
+                new CumQty(orderQty.Value));
 
             exReport.Set(clOrdID);
             exReport.Set(symbol);
             exReport.Set(orderQty);
-            exReport.Set(new LastQty(orderQty.getValue()));
-            exReport.Set(new LastPx(price.getValue()));
-            exReport.Set(new AvgPx(price.getValue()));
+            exReport.Set(new LastQty(orderQty.Value));
+            exReport.Set(new LastPx(price.Value));
+            exReport.Set(new AvgPx(price.Value));
 
             if (n.IsSetAccount())
                 exReport.SetField(n.Account);
@@ -375,7 +372,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX40.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX40.OrderCancelReject ocj = new QuickFix.FIX40.OrderCancelReject(new OrderID(orderid), msg.ClOrdID);
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
             ocj.Text = new Text("Executor does not support order cancels");
@@ -392,7 +389,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX41.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX41.OrderCancelReject ocj = new QuickFix.FIX41.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -410,7 +407,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX42.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX42.OrderCancelReject ocj = new QuickFix.FIX42.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -428,7 +425,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX43.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX43.OrderCancelReject ocj = new QuickFix.FIX43.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -446,7 +443,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX44.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX44.OrderCancelReject ocj = new QuickFix.FIX44.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.OTHER);
@@ -464,7 +461,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX50.OrderCancelRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX50.OrderCancelReject ocj = new QuickFix.FIX50.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.OTHER);
@@ -484,7 +481,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX40.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX40.OrderCancelReject ocj = new QuickFix.FIX40.OrderCancelReject(new OrderID(orderid), msg.ClOrdID);
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
             ocj.Text = new Text("Executor does not support order cancel/replaces");
@@ -501,7 +498,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX41.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX41.OrderCancelReject ocj = new QuickFix.FIX41.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -519,7 +516,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX42.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX42.OrderCancelReject ocj = new QuickFix.FIX42.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REPLACE_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -537,7 +534,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX43.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX43.OrderCancelReject ocj = new QuickFix.FIX43.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REPLACE_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.UNKNOWN_ORDER);
@@ -555,7 +552,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX44.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX44.OrderCancelReject ocj = new QuickFix.FIX44.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REPLACE_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.OTHER);
@@ -573,7 +570,7 @@ namespace Executor
 
         public void OnMessage(QuickFix.FIX50.OrderCancelReplaceRequest msg, SessionID s)
         {
-            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Obj : "unknown orderID";
+            string orderid = (msg.IsSetOrderID()) ? msg.OrderID.Value : "unknown orderID";
             QuickFix.FIX50.OrderCancelReject ocj = new QuickFix.FIX50.OrderCancelReject(
                 new OrderID(orderid), msg.ClOrdID, msg.OrigClOrdID, new OrdStatus(OrdStatus.REJECTED), new CxlRejResponseTo(CxlRejResponseTo.ORDER_CANCEL_REPLACE_REQUEST));
             ocj.CxlRejReason = new CxlRejReason(CxlRejReason.OTHER);