+++ QuickFIXn/Message/Message.cs
@@ -1,5 +1,4 @@
-﻿#nullable enable
-using System;
+﻿using System;
 using System.Text;
 using QuickFix.Fields;
 using System.Text.RegularExpressions;
@@ -146,7 +145,7 @@ namespace QuickFix
         public static string ExtractBeginString(string msgstr)
         {
             int i = 0;
-            return ExtractField(msgstr, ref i).Obj;
+            return ExtractField(msgstr, ref i).Value;
         }
 
         public static bool IsHeaderField(int tag)
@@ -205,6 +204,7 @@ namespace QuickFix
                     return false;
             }
         }
+
         public static bool IsTrailerField(int tag, DD? dd)
         {
             if (IsTrailerField(tag))
@@ -374,7 +374,7 @@ namespace QuickFix
 
                     if (Tags.MsgType.Equals(f.Tag))
                     {
-                        string msgType = f.Obj;
+                        string msgType = f.Value;
                         if (appDict is not null)
                         {
                             msgMap = appDict.GetMapForMessage(msgType);
@@ -454,7 +454,7 @@ namespace QuickFix
                         $"JSON message has invalid/missing beginString ({beginString}) and/or msgType ({msgType})");
                 }
 
-                IFieldMapSpec msgMap = appDict.GetMapForMessage(msgType);
+                IFieldMapSpec? msgMap = appDict.GetMapForMessage(msgType);
                 FromJson(document.RootElement.GetProperty("Header"),  beginString, msgType, msgMap, msgFactory, transportDict, Header);
                 FromJson(document.RootElement.GetProperty("Body"),    beginString, msgType, msgMap, msgFactory, appDict, this);
                 FromJson(document.RootElement.GetProperty("Trailer"), beginString, msgType, msgMap, msgFactory, transportDict, Trailer);
@@ -472,7 +472,7 @@ namespace QuickFix
         protected void FromJson(JsonElement jsonElement,
             string beginString,
             string msgType,
-            IFieldMapSpec msgMap,
+            IFieldMapSpec? msgMap,
             IMessageFactory? msgFactory,
             DD dataDict,
             FieldMap fieldMap)
@@ -654,48 +654,48 @@ namespace QuickFix
             {
                 SenderCompID senderCompId = new SenderCompID();
                 header.GetField(senderCompId);
-                if (senderCompId.Obj.Length > 0)
-                    Header.SetField(new TargetCompID(senderCompId.Obj));
+                if (senderCompId.Value.Length > 0)
+                    Header.SetField(new TargetCompID(senderCompId.Value));
             }
 
             if (header.IsSetField(Tags.SenderSubID))
             {
                 SenderSubID senderSubId = new SenderSubID();
                 header.GetField(senderSubId);
-                if (senderSubId.Obj.Length > 0)
-                    Header.SetField(new TargetSubID(senderSubId.Obj));
+                if (senderSubId.Value.Length > 0)
+                    Header.SetField(new TargetSubID(senderSubId.Value));
             }
 
             if (header.IsSetField(Tags.SenderLocationID))
             {
                 SenderLocationID senderLocationId = new SenderLocationID();
                 header.GetField(senderLocationId);
-                if (senderLocationId.Obj.Length > 0)
-                    Header.SetField(new TargetLocationID(senderLocationId.Obj));
+                if (senderLocationId.Value.Length > 0)
+                    Header.SetField(new TargetLocationID(senderLocationId.Value));
             }
 
             if (header.IsSetField(Tags.TargetCompID))
             {
                 TargetCompID targetCompId = new TargetCompID();
                 header.GetField(targetCompId);
-                if (targetCompId.Obj.Length > 0)
-                    Header.SetField(new SenderCompID(targetCompId.Obj));
+                if (targetCompId.Value.Length > 0)
+                    Header.SetField(new SenderCompID(targetCompId.Value));
             }
 
             if (header.IsSetField(Tags.TargetSubID))
             {
                 TargetSubID targetSubId = new TargetSubID();
                 header.GetField(targetSubId);
-                if (targetSubId.Obj.Length > 0)
-                    Header.SetField(new SenderSubID(targetSubId.Obj));
+                if (targetSubId.Value.Length > 0)
+                    Header.SetField(new SenderSubID(targetSubId.Value));
             }
 
             if (header.IsSetField(Tags.TargetLocationID))
             {
                 TargetLocationID targetLocationId = new TargetLocationID();
                 header.GetField(targetLocationId);
-                if (targetLocationId.Obj.Length > 0)
-                    Header.SetField(new SenderLocationID(targetLocationId.Obj));
+                if (targetLocationId.Value.Length > 0)
+                    Header.SetField(new SenderLocationID(targetLocationId.Value));
             }
 
             // optional routing tags
@@ -793,10 +793,15 @@ namespace QuickFix
                 m.Header.GetString(Tags.TargetCompID));
         }
 
-        private Object lock_ToString = new Object();
-        public override string ToString()
+        private readonly Object _lockConstructString = new Object();
+        /// <summary>
+        /// Update BodyLength in Header, update CheckSum in Trailer, and return a FIX string.
+        /// (This function changes the object state!)
+        /// </summary>
+        /// <returns></returns>
+        public string ConstructString()
         {
-            lock (lock_ToString)
+            lock (_lockConstructString)
             {
                 Header.SetField(new BodyLength(BodyLength()), true);
                 Trailer.SetField(new CheckSum(Fields.Converters.CheckSumConverter.Convert(CheckSum())), true);
@@ -805,12 +810,22 @@ namespace QuickFix
             }
         }
 
+        /// <summary>
+        /// Create a FIX-style string from the message.
+        /// This does NOT add/update BodyLength or CheckSum to the message header, or otherwise change the object state.
+        /// (Use <see cref="ConstructString"/> to compute and add/update BodyLength &amp; CheckSum.)
+        /// </summary>
+        /// <returns></returns>
+        public override string ToString() {
+            return Header.CalculateString() + CalculateString() + Trailer.CalculateString();
+        }
+
         protected int BodyLength()
         {
             return Header.CalculateLength() + CalculateLength() + Trailer.CalculateLength();
         }
 
-        private static string FieldMapToXML(DD? dd, FieldMap fields, int space)
+        private static string FieldMapToXml(DD? dd, FieldMap fields)
         {
             StringBuilder s = new StringBuilder();
 
@@ -833,7 +848,7 @@ namespace QuickFix
                 for (int counter = 1; counter <= fields.GroupCount(groupTag); counter++)
                 {
                     s.Append("<group>");
-                    s.Append(FieldMapToXML(dd, fields.GetGroup(counter, groupTag), space+1));
+                    s.Append(FieldMapToXml(dd, fields.GetGroup(counter, groupTag)));
                     s.Append("</group>");
                 }
             }
@@ -846,7 +861,7 @@ namespace QuickFix
         /// ToJSON() helper method.
         /// </summary>
         /// <returns>an XML string</returns>
-        private static StringBuilder FieldMapToJSON(StringBuilder sb, DD? dd, FieldMap fields, bool humanReadableValues)
+        private static StringBuilder FieldMapToJson(StringBuilder sb, DD? dd, FieldMap fields, bool humanReadableValues)
         {
             IList<int> numInGroupTagList = fields.GetGroupTags();
             IList<IField> numInGroupFieldList = new List<IField>();
@@ -890,16 +905,16 @@ namespace QuickFix
             foreach(IField numInGroupField in numInGroupFieldList)
             {
                 // The name of the NumInGroup field is the key of the JSON list containing the Group items
-                if (dd is not null && dd.FieldsByTag.ContainsKey(numInGroupField.Tag))
-                    sb.Append("\"" + dd.FieldsByTag[numInGroupField.Tag].Name + "\":[");
+                if (dd is not null && dd.FieldsByTag.TryGetValue(numInGroupField.Tag, out DDField? field))
+                    sb.Append("\"" + field.Name + "\":[");
                 else
                     sb.Append("\"" + numInGroupField.Tag + "\":[");
 
                 // Populate the JSON list with the Group items
                 for (int counter = 1; counter <= fields.GroupCount(numInGroupField.Tag); counter++)
                 {
-                    sb.Append("{");
-                    FieldMapToJSON(sb, dd, fields.GetGroup(counter, numInGroupField.Tag), humanReadableValues);
+                    sb.Append('{');
+                    FieldMapToJson(sb, dd, fields.GetGroup(counter, numInGroupField.Tag), humanReadableValues);
                     sb.Append("},");
                 }
 
@@ -927,13 +942,13 @@ namespace QuickFix
             StringBuilder s = new StringBuilder();
             s.Append("<message>");
             s.Append("<header>");
-            s.Append(FieldMapToXML(dataDictionary, Header, 4));
+            s.Append(FieldMapToXml(dataDictionary, Header));
             s.Append("</header>");
             s.Append("<body>");
-            s.Append(FieldMapToXML(dataDictionary, this, 4));
+            s.Append(FieldMapToXml(dataDictionary, this));
             s.Append("</body>");
             s.Append("<trailer>");
-            s.Append(FieldMapToXML(dataDictionary, Trailer, 4));
+            s.Append(FieldMapToXml(dataDictionary, Trailer));
             s.Append("</trailer>");
             s.Append("</message>");
             return s.ToString();
@@ -959,9 +974,9 @@ namespace QuickFix
             }
 
             StringBuilder sb = new StringBuilder().Append('{').Append("\"Header\":{");
-            FieldMapToJSON(sb, dataDictionary, Header, convertEnumsToDescriptions).Append("},\"Body\":{");
-            FieldMapToJSON(sb, dataDictionary, this, convertEnumsToDescriptions).Append("},\"Trailer\":{");
-            FieldMapToJSON(sb, dataDictionary, Trailer, convertEnumsToDescriptions).Append("}}");
+            FieldMapToJson(sb, dataDictionary, Header, convertEnumsToDescriptions).Append("},\"Body\":{");
+            FieldMapToJson(sb, dataDictionary, this, convertEnumsToDescriptions).Append("},\"Trailer\":{");
+            FieldMapToJson(sb, dataDictionary, Trailer, convertEnumsToDescriptions).Append("}}");
             return sb.ToString();
         }
     }