+++ QuickFIXn/Session.cs
@@ -1,4 +1,3 @@
-#nullable enable
 using System;
 using System.Collections.Generic;
 using System.Threading;
@@ -550,7 +549,7 @@ namespace QuickFix
                 if (_appDoesEarlyIntercept)
                     ((IApplicationExt)Application).FromEarlyIntercept(message, SessionID);
 
-                string msgType = msgBuilder.MsgType.Obj;
+                string msgType = msgBuilder.MsgType.Value;
                 string beginString = msgBuilder.BeginString;
 
                 if (!beginString.Equals(SessionID.BeginString))
@@ -569,7 +568,7 @@ namespace QuickFix
                 }
                 else
                 {
-                    SessionDataDictionary.Validate(message, beginString, msgType);
+                    DataDictionary.DataDictionary.Validate(message, SessionDataDictionary, SessionDataDictionary, beginString, msgType);
                 }
 
                 if (MsgType.LOGON.Equals(msgType))
@@ -600,7 +599,7 @@ namespace QuickFix
 
                 try
                 {
-                    if (MsgType.LOGON.Equals(msgBuilder.MsgType.Obj))
+                    if (MsgType.LOGON.Equals(msgBuilder.MsgType.Value))
                         Disconnect("Logon message is not valid");
                 }
                 catch (MessageParseError)
@@ -616,7 +615,7 @@ namespace QuickFix
             }
             catch (UnsupportedVersion uvx)
             {
-                if (MsgType.LOGOUT.Equals(msgBuilder.MsgType.Obj))
+                if (MsgType.LOGOUT.Equals(msgBuilder.MsgType.Value))
                 {
                     NextLogout(message!);
                 }
@@ -641,7 +640,7 @@ namespace QuickFix
                 }
                 else
                 {
-                    if (MsgType.LOGON.Equals(msgBuilder.MsgType.Obj))
+                    if (MsgType.LOGON.Equals(msgBuilder.MsgType.Value))
                     {
                         Log.OnEvent("Required field missing from logon");
                         Disconnect("Required field missing from logon");
@@ -789,7 +788,7 @@ namespace QuickFix
                             {
                                 GenerateSequenceReset(resendReq, begin, msgSeqNum);
                             }
-                            Send(msg.ToString());
+                            Send(msg.ConstructString());
                             begin = 0;
                         }
                         current = msgSeqNum + 1;
@@ -922,7 +921,7 @@ namespace QuickFix
                     DoTargetTooHigh(msg, msgSeqNum);
                     return false;
                 }
-                else if (checkTooLow && IsTargetTooLow(msgSeqNum))
+                if (checkTooLow && IsTargetTooLow(msgSeqNum))
                 {
                     DoTargetTooLow(msg, msgSeqNum);
                     return false;
@@ -1025,7 +1024,7 @@ namespace QuickFix
                 && SessionID.TargetCompID.Equals(senderCompId);
         }
 
-        /// FIXME - this fn always returns true-- should it ever be false?
+        /// TODO - this fn always returns true-- should it ever be false?
         protected bool IsTimeToGenerateLogon()
         {
             return true;
@@ -1151,11 +1150,9 @@ namespace QuickFix
                 Log.OnEvent("Sent ResendRequest FROM: " + startSeqNum + " TO: " + endSeqNum);
                 return true;
             }
-            else
-            {
-                Log.OnEvent("Error sending ResendRequest (" + startSeqNum + " ," + endSeqNum + ")");
-                return false;
-            }
+
+            Log.OnEvent("Error sending ResendRequest (" + startSeqNum + " ," + endSeqNum + ")");
+            return false;
         }
 
         protected void GenerateResendRequest(string beginString, SeqNumType msgSeqNum)
@@ -1336,8 +1333,10 @@ namespace QuickFix
                     msgSeqNum = message.Header.GetULong(Fields.Tags.MsgSeqNum);
                     reject.SetField(new Fields.RefSeqNum(msgSeqNum));
                 }
-                catch (Exception)
-                { }
+                catch (Exception ex)
+                {
+                    Log.OnEvent($"Exception while setting RefSeqNum: {ex}");
+                }
             }
 
             if (string.CompareOrdinal(beginString, FixValues.BeginString.FIX42) >= 0)
@@ -1517,7 +1516,7 @@ namespace QuickFix
                 }
                 else
                 {
-                    NextMessage(msg.ToString());
+                    NextMessage(msg.ConstructString());
                 }
                 return true;
             }
@@ -1547,12 +1546,12 @@ namespace QuickFix
                         Fields.ResetSeqNumFlag resetSeqNumFlag = new Fields.ResetSeqNumFlag(false);
                         if (message.IsSetField(resetSeqNumFlag))
                             message.GetField(resetSeqNumFlag);
-                        if (resetSeqNumFlag.getValue())
+                        if (resetSeqNumFlag.Value)
                         {
                             _state.Reset("ResetSeqNumFlag");
                             message.Header.SetField(new Fields.MsgSeqNum(_state.NextSenderMsgSeqNum));
                         }
-                        _state.SentReset = resetSeqNumFlag.Obj;
+                        _state.SentReset = resetSeqNumFlag.Value;
                     }
                 }
                 else
@@ -1567,7 +1566,7 @@ namespace QuickFix
                     }
                 }
 
-                string messageString = message.ToString();
+                string messageString = message.ConstructString();
                 if (0 == seqNum)
                     Persist(message, messageString);
                 return Send(messageString);